Add relayer workflow and enhance requester workflow
The code updates include the addition of a relayer workflow, enhancement of the requester workflow, and adding session tolerance in the RPC configuration. Moreover, we further updated the MongoDB related structures and methods, with necessary changes reflected in corresponding test files. The modifications also accommodate some changes in the representation of data, depicted by modifications in various Go structures.